+++ remoting/base/shared_buffer.h
@@ -27,12 +27,12 @@ class SharedBuffer
   // Opens an existing shared memory buffer and maps it to the memory of
   // the calling process (in read only mode). If the operation fails for any
   // reason, ptr() method will return NULL.
-  SharedBuffer(intptr_t id, base::SharedMemoryHandle handle, uint32 size);
+  SharedBuffer(int id, base::SharedMemoryHandle handle, uint32 size);
 
   // Opens an existing shared memory buffer created by a different process and
   // maps it to the memory of the calling process (in read only mode). If
   // the operation fails for any reason, ptr() method will return NULL.
-  SharedBuffer(intptr_t id, base::SharedMemoryHandle handle,
+  SharedBuffer(int id, base::SharedMemoryHandle handle,
                base::ProcessHandle process, uint32 size);
 
   // Returns pointer to the beginning of the allocated data buffer. Returns NULL
@@ -43,17 +43,17 @@ class SharedBuffer
   // data buffer.
   base::SharedMemoryHandle handle() const { return shared_memory_.handle(); }
 
-  intptr_t id() const { return id_; }
+  int id() const { return id_; }
   uint32 size() const { return size_; }
 
-  void set_id(intptr_t id) { id_ = id; }
+  void set_id(int id) { id_ = id; }
 
  private:
   friend class base::RefCountedThreadSafe<SharedBuffer>;
   virtual ~SharedBuffer();
 
   // Unique identifier of the buffer or 0 if ID hasn't been set.
-  intptr_t id_;
+  int id_;
 
   // Shared memory section backing up the buffer.
   base::SharedMemory shared_memory_;
